Move networking and table view to their own files
